+++ Vault-Coder/src/utopia/vault/coder/model/scala/code/CodeLine.scala
@@ -22,7 +22,8 @@ object CodeLine
 	private lazy val repeatableRegexes = Vector(
 		Regex("with ") -> false,
 		(Regex.escape(',') + Regex.whiteSpace.noneOrOnce).withinParenthesis.ignoringQuotations -> true,
-		Regex.escape('\"') + (!Regex.escape('\"')).zeroOrMoreTimes + Regex.escape('\"') -> false,
+		Regex("s").noneOrOnce + Regex.escape('\"') + (!Regex.escape('\"')).zeroOrMoreTimes +
+			Regex.escape('\"') -> false,
 		(Regex.anyOf("+-*/").oneOrMoreTimes + Regex.whiteSpace) -> true,
 		(Regex.whiteSpace + Regex.word + Regex.whiteSpace) -> true
 	)